Support setting baudrate. enableGPRS more robust
enableGPRS will return true if it's already on, instead of returning
false.
Setting a baudrate is more reliable, so I've added the ability to use
it. Specifically, at startup we should check if there's a "RDY"
unsolicited command that is sent. If so, there's a set shared baudrate.
   - Note the caveat: If you set the baudrate of the FONA and then
     change the baudrate you are communicating to the FONA with, nothing
     will work (specifically, autobauding won't work).
I've also removed a 7 second startup wait. I'm not sure why it exists.