Make enableGPRS more reliable
I was checking CGATT=1 before, but that's not enough to tell that the
GPRS is on or not. So instead of checking if it's on, I just make sure
to turn it off and then turn it back on.

I have to turn it off because otherwise SAPBR=1,1 will return an error
if the GPRS is already on. To keep it from returning an error, I turn it
off.