Fugly support for ringing URCs
This supports following URCs:
- RING
- +CLIP...

It's in two different places. It's in the new "readring" method that is
used when I know the URC's are coming (by noticing the RI interrupt).
The modification of `readline` exists because I may get URC's in the
middle of a readline. I can't read the RI pin because it may be out of
date from the serial. I *could* check if the interrupt has been hit and
a variable has not been cleared. In which case I'd have to stop
`readline` and jump over to `readring`, or something like that